+++ kernel/src/acpi/mod.rs
@@ -24,7 +24,7 @@ pub struct ACPITables {
 
 impl ACPITables {
     // Should read this straight from the boot info (UEFI/BIOS)
-    unsafe fn read_acpi_tables(raw_xsdp_physical_address: u64) -> ACPITables {
+    unsafe fn read_acpi_tables(raw_xsdp_physical_address: usize) -> ACPITables {
         let xsdp = XSDP::init(raw_xsdp_physical_address);
         let xsdt = XSDT::read_from_raw_address(xsdp.xsdt_address);
         let raw_madt_physical_address = xsdt.try_get_raw_sdt_table_address(&SDTSignature::MADT).unwrap();
@@ -33,7 +33,7 @@ impl ACPITables {
     }
 }
 
-pub fn read_acpi_tables(raw_xsdp_physical_address: u64) {
+pub fn read_acpi_tables(raw_xsdp_physical_address: usize) {
     let acpi_tables = unsafe { ACPITables::read_acpi_tables(raw_xsdp_physical_address) };
     ACPI_TABLES.get_or_init(move || acpi_tables);
     log::info!("Successfully parsed ACPI tables");
