+++ kernel/src/acpi/xsdt.rs
@@ -6,20 +6,20 @@ use crate::acpi::sdt::{
     SDTSignature,
     SystemDescriptorTable,
 };
-use crate::mmu::virtual_address::VirtualAddress;
+use crate::mmu::address::VirtualAddress;
 
 #[derive(Debug)]
 #[repr(C, packed)]
 pub struct XSDT {
     header: SDTHeader,
-    sdt_address_table: &'static [u64],
+    sdt_address_table: &'static [usize],
 }
 
 impl SystemDescriptorTable for XSDT {
-    unsafe fn read_from_raw_address(raw_xsdt_physical_address: u64) -> Self {
+    unsafe fn read_from_raw_address(raw_xsdt_physical_address: usize) -> Self {
         let header = SDTHeader::try_read_from_phys_addr(raw_xsdt_physical_address, &SDTSignature::XSDT).unwrap();
-        let size_of_sdt_header = size_of::<SDTHeader>() as u64;
-        let sdt_address_table_size = header.length as u64 - size_of_sdt_header;
+        let size_of_sdt_header = size_of::<SDTHeader>() as usize;
+        let sdt_address_table_size = header.length as usize - size_of_sdt_header;
         let sdt_address_table_length = (sdt_address_table_size / 8) as usize;
         let sdt_address_table_virt_addr =
             VirtualAddress::with_kernel_base_offset(raw_xsdt_physical_address + size_of_sdt_header);
@@ -32,7 +32,7 @@ impl SystemDescriptorTable for XSDT {
 }
 
 impl XSDT {
-    pub unsafe fn try_get_raw_sdt_table_address(&self, sdt_signature: &SDTSignature) -> Option<u64> {
+    pub unsafe fn try_get_raw_sdt_table_address(&self, sdt_signature: &SDTSignature) -> Option<usize> {
         for raw_sdt_address in self.sdt_address_table {
             if let Ok(_) = SDTHeader::try_read_from_phys_addr(*raw_sdt_address, sdt_signature) {
                 return Some(*raw_sdt_address);
